requirements.txt
Django==1.6
Pillow==2.2.1
South==0.8.4
argparse==1.2.1
distribute==0.6.49
gunicorn==0.17.2
sorl-thumbnail==11.12
wsgiref==0.1.2
